lib: compare a counter and a composite's size in one signedness
A composite carries its element count in word, and large-object doubles
that type: an unsigned short in the default build, an unsigned int in
the wide one. Only the first of those promotes to int, so a comparison
between a signed counter and a count is a comparison of two signed
values in the default build and a comparison across signednesses in the
wide one -- the same line, two meanings, and a diagnostic in only the
build almost nobody runs. Forty-five such comparisons had collected
there, none of them visible from the default build.

Each is settled in one of two ways rather than by silencing it. A
counter that indexes a composite and does nothing else takes the type of
the bound it is compared against, so that the two are the same kind of
thing and the loop cannot describe a position the composite could not
have. The dictionary walks in the collector and its verifier, the array
and file operators that step a string or an array through, the polygon
walk in the X device and the raster clear all take that form.

The rest are not counters. A device coordinate arrives from anywhere on
the page and is signed on purpose; a length answered by a writer, a read
or a write is a report that may fall short; a restore's depth is counted
at the moment it is asked for. Narrowing a size into those would make a
negative answer compare as an enormous positive one and inverts the very
test that is meant to catch it, so the size is widened into integer
instead, which is strictly wider than word in both builds and therefore
loses nothing, and the comparison stays signed and keeps its meaning.
Two comparisons already carried a cast the other way and are corrected
here for the same reason.

The array and dictionary constructors kept their counted length in an
int while comparing it against the widest size the field can hold, which
in the wide build is a bound an int cannot reach: the guard was
permanently false and the length was truncated on its way in from an
operand that is already wider. Both now hold that length in the type the
operand arrives in, which makes the guard mean something again.

Commit message generated by Claude (Anthropic).
Source code generated by Claude (Anthropic).

+++ src/lib/xpost_op_misc.c
@@ -108,7 +108,7 @@ Xpost_Object bind(Xpost_Context *ctx,
     }
     seen->ents[seen->n++] = ent;
 
-    for (i = 0; i < p.comp_.sz; i++)
+    for (i = 0; i < (integer)p.comp_.sz; i++)
     {
         t = xpost_array_get(ctx, p, i);
         switch(xpost_object_get_type(t))
